Modifications to allow handling of source grids that are larger than Integer.MAX_VALUE in size.  Storage allocated by PixelMap is reduced by using resizeable arrays (RArrays) with appropriate underlying data types.
git-svn-id: https://ncwms.svn.sourceforge.net/svnroot/ncwms/trunk@938 f020452a-5e1f-0410-8fa4-ae2b6fb27d5e

+++ src/java/uk/ac/rdg/resc/edal/cdm/CdmUtils.java
@@ -439,7 +439,9 @@ public final class CdmUtils
         if (zIndices == null) zIndices = Arrays.asList(-1);
         List<List<Float>> data = new ArrayList<List<Float>>(zIndices.size());
         for (int zIndex : zIndices) {
-            data.add(readHorizontalPoints(nc, grid, tIndex, zIndex, pixelMap, targetDomain.size()));
+            // It's very unlikely that the target domain will be bigger than
+            // Integer.MAX_VALUE
+            data.add(readHorizontalPoints(nc, grid, tIndex, zIndex, pixelMap, (int)targetDomain.size()));
         }
         return data;
     }
@@ -509,10 +511,12 @@ public final class CdmUtils
         {
             // There is no overlap between the source data grid and the target
             // domain.  Return a list of null values.
-            return nullList(targetDomain.size());
+            // It's very unlikely that the target domain will be bigger than
+            // Integer.MAX_VALUE
+            return nullList((int)targetDomain.size());
         }
 
-        return readHorizontalPoints(nc, grid, tIndex, zIndex, pixelMap, targetDomain.size());
+        return readHorizontalPoints(nc, grid, tIndex, zIndex, pixelMap, (int)targetDomain.size());
     }
 
     static List<Float> readHorizontalPoints(NetcdfDataset nc, GridDatatype grid,
